requirements.txt
asgiref==3.2.7
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
django-ajax-selects==1.9.1
django-allauth==0.41.0
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-elasticsearch-dsl==7.1.1
django-js-asset==1.2.2
django-mptt-admin==1.0.1
django-mptt==0.11.0
django-parler==2.0.1
django-smart-selects==1.5.4
Django==3.0.5
elasticsearch-dsl==7.1.0
elasticsearch==7.6.0
gunicorn==20.0.4
idna==2.9
ipython-genutils==0.2.0
ipython==7.13.0
jedi==0.16.0
mysqlclient==1.3.10
oauthlib==3.1.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
progressbar2==3.50.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python-utils==2.4.0
python3-openid==3.1.0
pytz==2019.3
requests-oauthlib==1.3.0
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.8
wcwidth==0.1.9
